refactor: model -> domain 폴더로 변경
